first chart done
