+++ src/index.html
@@ -22,8 +22,9 @@
 
   <!-- ChartJS Component -->
   <div>
-  <canvas id="myChart" width="400" height="400"></canvas>
-  </div>  
+    <canvas id="myChart" width="400" height="400"></canvas>
+  </div>
+  <!-- end ChartJS -->
 
   <div id="parentElement">
 
@@ -76,7 +77,7 @@
         </div>
         <div class="modal-footer">
           <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-          <button type="button" class="btn btn-primary" data-dismiss="modal"id="btn--newCategory">Create New Category</button>
+          <button type="button" class="btn btn-primary" data-dismiss="modal" id="btn--newCategory">Create New Category</button>
         </div>
       </div>
     </div>
