+++ src/styles/_04defaults.scss
@@ -148,9 +148,9 @@ img{
     border: 1px solid $base-text-color;
 }
 
-:empty { 
-    background-color: $test-color;
-}
+// :empty { 
+//     background-color: $test-color;
+// }
 
 .column {
     height: 100%;
